Move network error handling into use-case
